* SimpleAdapter and Uitls is moved to UI module * added more documentation to SimpleAdapter.kt
